# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MVTable 后端服务 API",
        "title": "MVTable",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MVTable",
            "email": "mvtable@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/team/add-member": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "向团队添加新成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "添加团队成员",
                "parameters": [
                    {
                        "description": "添加成员请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/batch-get-users": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据用户ID列表批量查询用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "批量查询用户",
                "parameters": [
                    {
                        "description": "批量查询用户请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BatchGetUsersReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.BatchGetUsersRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/team/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建新的团队",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "创建团队",
                "parameters": [
                    {
                        "description": "创建团队请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateTeamReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/delete": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除团队",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "删除团队",
                "parameters": [
                    {
                        "description": "删除团队请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeleteTeamReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/delete-member": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "从团队中删除成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "删除团队成员",
                "parameters": [
                    {
                        "description": "删除成员请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mvtable_internal_app_team_model.DeleteMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取当前用户的团队列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "获取团队列表",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ListTeamReq"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/team/member-list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取团队成员列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "获取团队成员列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "团队ID",
                        "name": "teamId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "searchWord",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.GetMemberListRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/team/search-user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据邮箱或昵称搜索用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "搜索用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "searchWord",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "搜索成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SearchUserRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/team/update": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新团队信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "更新团队",
                "parameters": [
                    {
                        "description": "更新团队请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateTeamReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/v1/team/update-member": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新团队成员信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "团队"
                ],
                "summary": "更新团队成员",
                "parameters": [
                    {
                        "description": "更新成员请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mvtable_internal_app_team_model.UpdateMemberReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "constants.TeamIdentity": {
            "type": "string",
            "enum": [
                "creator",
                "manager",
                "member"
            ],
            "x-enum-varnames": [
                "TeamIdentityCreator",
                "TeamIdentityManager",
                "TeamIdentityMember"
            ]
        },
...
        "middleware.Response": {
            "description": "统一响应结构",
            "type": "object",
            "properties": {
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string"
                },
                "state": {
                    "description": "响应状态码",
                    "type": "integer"
                }
            }
        },
        "model.AddMemberReq": {
            "type": "object",
            "required": [
                "identity",
                "teamId",
                "userId"
            ],
            "properties": {
                "identity": {
                    "$ref": "#/definitions/constants.TeamIdentity"
                },
                "teamId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "model.BatchGetUsersReq": {
            "type": "object",
            "required": [
                "userIds"
            ],
            "properties": {
                "userIds": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.BatchGetUsersRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserBasicInfo"
                    }
                }
            }
        },
...
        "model.CreateTeamReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.DeleteTeamReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "model.GetMemberListRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TeamMember"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.ListTeamReq": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TeamListItem"
                    }
                }
            }
        },
...
        "model.SearchUserRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserBasicInfo"
                    }
                }
            }
        },
...
        "model.TeamListItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayAvatarList": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "identity": {
                    "$ref": "#/definitions/constants.TeamIdentity"
                },
                "logo": {
                    "type": "string"
                },
                "memberCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.TeamMember": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "identity": {
                    "$ref": "#/definitions/constants.TeamIdentity"
                },
                "nickname": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateTeamReq": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.UserBasicInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "mvtable_internal_app_team_model.DeleteMemberReq": {
            "type": "object",
            "required": [
                "teamId",
                "userId"
            ],
            "properties": {
                "teamId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "mvtable_internal_app_team_model.UpdateMemberReq": {
            "type": "object",
            "required": [
                "identity",
                "teamId",
                "userId"
            ],
            "properties": {
                "identity": {
                    "$ref": "#/definitions/constants.TeamIdentity"
                },
                "teamId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "输入Bearer token进行认证，格式: Bearer {token}。例如: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}